# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Financial Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Financial Management application, providing endpoints for managing accounts, transactions, and budgets."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/budgets": {
      "get": {
        "summary": "Retrieve all budgets",
        "description": "Get a list of all spending budgets from the database",
        "tags": [
          "Budgets"
        ],
        "responses": {
          "200": {
            "description": "Budgets retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Budgets retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Budget"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new budget",
        "tags": [
          "Budgets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Budget created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Budget created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Budget"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/budgets/{id}": {
      "get": {
        "summary": "Get budget by ID",
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Budget ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Budget retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Budget retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Budget"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Budget not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update budget by ID",
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Budget ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Budget updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Budget updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Budget not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete budget by ID",
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Budget ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Budget deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Budget deleted successfully"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Budget not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Budget": {
        "type": "object",
        "required": [
          "userId",
          "category",
          "limit",
          "spent",
          "month",
          "year"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the budget"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who owns the budget"
          },
          "category": {
            "type": "string",
            "description": "Budget category",
            "example": "Food"
          },
          "limit": {
            "type": "number",
            "description": "Budget limit amount",
            "example": 500
          },
          "spent": {
            "type": "number",
            "description": "Amount already spent",
            "example": 150
          },
          "month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "description": "Month of the budget (1-12)",
            "example": 10
          },
          "year": {
            "type": "integer",
            "minimum": 2000,
            "description": "Year of the budget",
            "example": 2023
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "example": {
          "id": "budget_123",
          "userId": "user_456",
          "category": "Food",
          "limit": 500,
          "spent": 150,
          "month": 10,
          "year": 2023,
          "createdAt": "2023-10-01T10:00:00.000Z",
          "updatedAt": "2023-10-01T10:00:00.000Z"
        }
      },
      "CreateBudgetRequest": {
        "type": "object",
        "required": [
          "category",
          "limit",
          "spent",
          "month",
          "year"
        ],
        "properties": {
          "category": {
            "type": "string",
            "description": "Budget category",
            "example": "Food"
          },
          "limit": {
            "type": "number",
            "description": "Budget limit amount",
            "example": 500
          },
          "spent": {
            "type": "number",
            "description": "Initial amount spent",
            "example": 0
          },
          "month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "description": "Month of the budget (1-12)",
            "example": 10
          },
          "year": {
            "type": "integer",
            "minimum": 2000,
            "description": "Year of the budget",
            "example": 2023
          }
        },
        "example": {
          "category": "Food",
          "limit": 500,
          "spent": 0,
          "month": 10,
          "year": 2023
        }
      },
      "UpdateBudgetRequest": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Budget category",
            "example": "Groceries"
          },
          "limit": {
            "type": "number",
            "description": "Budget limit amount",
            "example": 600
          },
          "spent": {
            "type": "number",
            "description": "Amount already spent",
            "example": 200
          },
          "month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "description": "Month of the budget (1-12)",
            "example": 10
          },
          "year": {
            "type": "integer",
            "minimum": 2000,
            "description": "Year of the budget",
            "example": 2023
          }
        },
        "example": {
          "limit": 600,
          "spent": 200
        }
      },
...
    }
  },
...
}